[DVK-64] docs: 알림 API에 Swagger 적용

+++ src/main/kotlin/com/devooks/backend/notification/v1/application/port/in/GetNotificationUseCase.kt
@@ -6,6 +6,6 @@ import java.util.*
 import org.springframework.data.domain.Page
 
 interface GetNotificationUseCase {
-    suspend fun getCountOfUnchecked(memberId: UUID): Long
+    suspend fun getCountOfUnchecked(memberId: UUID): Int
     suspend fun getNotifications(request: GetNotificationsRequest): Page<NotificationResponse>
 }
